:sparkles: `[platform]` Add way to run commands as a user with privilege
