+++ changes/20231013112742.feature
@@ -1,0 +1,1 @@
+:sparkles: `[platform]` Add way to run commands as a user with privileges on posix systems
